+++ gsy-ai-cloud/ai-pojo/src/main/java/com/guanshiyun/chat/ChatRecord.java
@@ -5,7 +5,7 @@ import lombok.experimental.FieldNameConstants;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
 
-import java.math.BigInteger;
+
 import java.time.LocalDateTime;
 
 /**
@@ -22,18 +22,18 @@ import java.time.LocalDateTime;
 public class ChatRecord {
     //聊天记录id
     @Id
-    private BigInteger id;
+    private Long id;
     //标题
     private String title;
     /**
      * 创建者，目前使用 SysUser 的 id 编号
      *
      */
-    public BigInteger creator;
+    public Long creator;
     /**
      * 更新者，目前使用 SysUser 的 id 编号
      */
-    public BigInteger updater;
+    public Long updater;
     /**
      * 创建时间
      */
