+++ gsy-ai-cloud/ai-pojo/src/main/java/com/guanshiyun/mymongodb/ChatRecordContent.java
@@ -6,7 +6,7 @@ import lombok.experimental.FieldNameConstants;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.math.BigInteger;
+
 import java.time.LocalDateTime;
 import java.util.List;
 /**
@@ -21,24 +21,24 @@ import java.util.List;
 public class ChatRecordContent {
     //聊天记录内容id
     @Id
-    private BigInteger id;
+    private Long id;
     // 会话标题
     private String title;
     //发送方id
-    private BigInteger senderId;
+    private Long senderId;
     //接收方id
-    private BigInteger receiverId;
+    private Long receiverId;
     //内容
     private List<ContentText> contentTexts;
     /**
      * 创建者，目前使用 SysUser 的 id 编号
      *
      */
-    public BigInteger creator;
+    public Long creator;
     /**
      * 更新者，目前使用 SysUser 的 id 编号
      */
-    public BigInteger updater;
+    public Long updater;
     /**
      * 创建时间
      */
