+++ gsy-ai-cloud/ai-server/src/main/java/com/guanshiyun/controller/chat/vo/ChatRecordSaveVO.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 import lombok.experimental.FieldNameConstants;
 
-import java.math.BigInteger;
+
 import java.time.LocalDateTime;
 
 /**
@@ -23,18 +23,18 @@ import java.time.LocalDateTime;
 @Accessors(chain = true)
 public class ChatRecordSaveVO {
     //聊天记录id
-    private BigInteger id;
+    private Long id;
     //标题
     private String title;
     /**
      * 创建者，目前使用 SysUser 的 id 编号
      *
      */
-    public BigInteger creator;
+    public Long creator;
     /**
      * 更新者，目前使用 SysUser 的 id 编号
      */
-    public BigInteger updater;
+    public Long updater;
     /**
      * 创建时间
      */
