+++ gsy-ai-cloud/ai-server/src/main/java/com/guanshiyun/repository/knowledge/KnowledgeRepositorySql.java
@@ -3,7 +3,7 @@ package com.guanshiyun.repository.knowledge;
 import com.guanshiyun.knowledge.KnowledgeRepository;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-import java.math.BigInteger;
 
-public interface KnowledgeRepositorySql extends ReactiveCrudRepository<KnowledgeRepository, BigInteger> {
+
+public interface KnowledgeRepositorySql extends ReactiveCrudRepository<KnowledgeRepository, Long> {
 }
